refactor(Player関連): Youtubeのディスプレイとコントローラを分割
ユーザの細かい操作も同期できるようにYoutubeの動画再生ディスプレイと操作コントローラを分割。
